Langchain Documentation Helper/requirement.txt
pinecone-client
langchain
langchain-openai
langchain-community
streamlit
google-generativeai
python-dotenv
langchain-google-genai
openai
black
PyPDF2
pypdf
beautifulsoup4